Improve and cleanup code in runelite-client (#3859)
* Remove redundant escape char
* Remove double negation
* Replace while loop with forEach
* Replace collections.sort with list.sort as its faster
* Replace if/elseif chain with switch
* Remove redundant call to format()
* Use bulk operation instead of iteration
* Remove unnecessary contains check
* Remove redundant type casts

+++ runelite-client/src/main/java/net/runelite/client/plugins/instancemap/InstanceMapOverlay.java
@@ -200,8 +200,8 @@ class InstanceMapOverlay extends Overlay
 		int tileX = playerLoc.getRegionX();
 		int tileY = (tiles[0].length - 1) - playerLoc.getRegionY(); // flip the y value
 
-		int x = (int) (tileX * TILE_SIZE);
-		int y = (int) (tileY * TILE_SIZE);
+		int x = tileX * TILE_SIZE;
+		int y = tileY * TILE_SIZE;
 		graphics.setColor(dotColor);
 		graphics.fillRect(x, y, PLAYER_MARKER_SIZE, PLAYER_MARKER_SIZE);//draw the players point on the map
 		graphics.setColor(outlineColor);
